include time in default end date
The date field that is queried is a DateTime field, so without this
change the report is not inclusive of the end date. This change
adds in the time in the end date so this is more obvious to the user
that the time is needed to get the end date's tickets in the report.